--- OilPrice-Where/Cell/Setting/AboutUsTableViewCell.swift
@@ -1,15 +1,0 @@
-//
-//  AboutUsTableViewCell.swift
-//  OilPrice-Where
-//
-//  Created by 박상욱 on 2020/07/28.
-//  Copyright © 2020 sangwook park. All rights reserved.
-//
-
-import UIKit
-
-class AboutUsTableViewCell: UITableViewCell {
-   static let identifier = "AboutUsTableViewCell"
-   @IBOutlet weak var nameLabel: UILabel!
-   @IBOutlet weak var linkLabel: UILabel!
-}
